Add (extremely) rudimentary kernel based on Inline::Python.
Currently, this is little more than a proof of concept. It does little more
than stringify the return value from a given piece of code. It makes
absolutely no attempt at handling exceptions, and it may fail intermittently
(I've had some unexplained failures).

To get it to work, you need to edit inline-wrapper/kernel.json so that the
path to kernel.p6 is correct. Then:
    ln -s ~/path/to/inline-wrapper ~/.local/share/jupyter/kernels/perl6